requirements.txt
Django==2.0
django-solo==1.1.3
pytz==2018.3
